Reorganizing into common suites to make future model source  development easier
