+++ module/model/extension/base.test.ts
@@ -1,6 +1,7 @@
 import { SchemaRegistry } from '@travetto/schema';
 import { DependencyRegistry } from '@travetto/di';
 import { Class } from '@travetto/registry';
+import { AfterEach, AfterAll } from '@travetto/test';
 
 import { ModelRegistry } from '../src/registry';
 import { ModelSource } from '../src/service/source';
@@ -8,6 +9,7 @@ import { ModelSource } from '../src/service/source';
 export abstract class BaseModelTest {
 
   abstract get configClass(): Class<{ namespace: string }>;
+  abstract get sourceClass(): Class<ModelSource>;
 
   async init() {
     await SchemaRegistry.init();
@@ -17,12 +19,14 @@ export abstract class BaseModelTest {
     await ModelRegistry.init();
   }
 
+  @AfterAll()
   async clear() {
     const mms = (await DependencyRegistry.getInstance(ModelSource));
     await mms.clearDatabase();
 
   }
 
+  @AfterEach()
   async reinit() {
     const mms = (await DependencyRegistry.getInstance(ModelSource));
     await mms.clearDatabase();
